+++ LBEngine/src/CollisionComponent.cpp
@@ -3,11 +3,11 @@
 
 using namespace DirectX::SimpleMath;
 
-CollisionComponent::CollisionComponent(std::function<void(Entity*, Entity*, DWORD)> func, double upleftX, double upleftY, double downrightX, double downrightY) : CollisionComponent(func, Vector2(upleftX, upleftY), Vector2(downrightX, downrightY))
+CollisionComponent::CollisionComponent(std::function<void(Entity*, Entity*, double)> func, double upleftX, double upleftY, double downrightX, double downrightY) : CollisionComponent(func, Vector2(upleftX, upleftY), Vector2(downrightX, downrightY))
 {}
 
 
-CollisionComponent::CollisionComponent(std::function<void(Entity*, Entity*, DWORD)> func, DirectX::SimpleMath::Vector2 upleft, DirectX::SimpleMath::Vector2 downright)
+CollisionComponent::CollisionComponent(std::function<void(Entity*, Entity*, double)> func, DirectX::SimpleMath::Vector2 upleft, DirectX::SimpleMath::Vector2 downright)
 {
 	m_upleft = upleft;
 	m_downright = downright;
